feat: ✨ Organize Document Classes
